requirements.txt
python-telegram-bot==20.7
LunarDate==0.2.2
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
aiohttp==3.9.1
